add chave
