+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@material-ui/core": "^4.11.1",
     "@material-ui/icons": "^4.9.1",
+    "md5": "^2.3.0",
     "mongodb": "^3.5.9",
     "next": "latest",
     "react": "^16.13.1",
