+++ src/pages/_app.js
@@ -1,11 +1,20 @@
-import '../styles/index.scss'
+import "../styles/index.scss";
+
+import { MuiThemeProvider, createMuiTheme } from "@material-ui/core/styles";
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      main: "#0070f3",
+    },
+  },
+});
 
 function MyApp({ Component, pageProps }) {
   return (
-    <React.Fragment>
+    <MuiThemeProvider theme={theme}>
       <Component {...pageProps} />
-    </React.Fragment>
-  )
+    </MuiThemeProvider>
+  );
 }
 
 export default MyApp;
